requirements.txt
requests~=2.32.4
python-dotenv~=1.1.0
